rework task get, creation, update and delete to use assigned_to rework

+++ src/main/java/ch/cern/todo/service/TaskService.java
@@ -7,6 +7,7 @@ import ch.cern.todo.model.business.Task;
 import ch.cern.todo.model.mapper.TaskMapper;
 import ch.cern.todo.repository.TaskCategoriesRepository;
 import ch.cern.todo.repository.TaskRepository;
+import ch.cern.todo.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -24,6 +25,7 @@ public class TaskService {
     private final TaskMapper taskMapper;
     private final TaskRepository taskRepository;
     private final TaskCategoriesRepository taskCategoriesRepository;
+    private final UserRepository userRepository;
 
     public Task getTask(Long taskId, LoggedUserInfo loggedUserInfo) {
         val taskEntity = taskRepository.getTaskById(taskId);
@@ -33,7 +35,7 @@ public class TaskService {
         // TODO: If security is a concern and we want to avoid exposing unnecessary information,
         //  the exception should be a generic 404. In that case, loggedUserInfo.userId() could be added as
         //  a filter of getTaskById() paramethers
-        if (!loggedUserInfo.isAdmin() && !taskEntity.getCreationUserId().equals(loggedUserInfo.userId())) {
+        if (!loggedUserInfo.isAdmin() && !taskEntity.getAssignedToUserId().equals(loggedUserInfo.userId())) {
             throw new RuntimeException("You are not allowed to see this task"); // TODO: custom exception
         }
 
@@ -48,14 +50,18 @@ public class TaskService {
 
     @Transactional
     public Task createTask(Task task, LoggedUserInfo loggedUserInfo) {
+        // TODO: strange behaviour due to the lack of detailed specification:
+        //  a user can create a task for another user, but after that, if he is not admin, he will not be able to see it
+
         val taskCategory = taskCategoriesRepository.getReferenceById(task.getCategoryId());
+        val user = userRepository.getReferenceById(task.getAssignedToUserInfo().getUserId());
 
         val taskEntity = taskMapper.businessToEntity(task);
         taskEntity.setTaskCategory(taskCategory);
         taskEntity.setCreationUserId(loggedUserInfo.userId());
-        taskEntity.setCreationUsername(loggedUserInfo.username());
         taskEntity.setUpdateUserId(loggedUserInfo.userId());
-        taskEntity.setUpdateUsername(loggedUserInfo.username());
+        taskEntity.setAssignedToUserId(user.getUserId());
+        taskEntity.setAssignedToUsername(user.getUsername());
 
         taskRepository.save(taskEntity);
 
@@ -66,7 +72,7 @@ public class TaskService {
     public Task updateTask(Task task, LoggedUserInfo loggedUserInfo) {
         val taskEntity = taskRepository.getReferenceById(task.getTaskId());
 
-        if (!loggedUserInfo.isAdmin() && !taskEntity.getCreationUserId().equals(loggedUserInfo.userId())) {
+        if (!canModifyTask(taskEntity.getCreationUserId(), loggedUserInfo)) {
             throw new RuntimeException("You are not allowed to update this task"); // TODO: custom exception
         }
 
@@ -86,8 +92,14 @@ public class TaskService {
             taskEntity.setTaskCategory(taskCategory);
         }
 
+        if (task.getAssignedToUserInfo().getUserId() != null
+            && !task.getAssignedToUserInfo().getUserId().equals(taskEntity.getAssignedToUserId())) {
+            val userEntity = userRepository.getReferenceById(task.getAssignedToUserInfo().getUserId());
+            taskEntity.setAssignedToUserId(userEntity.getUserId());
+            taskEntity.setAssignedToUsername(userEntity.getUsername());
+        }
+
         taskEntity.setUpdateUserId(loggedUserInfo.userId());
-        taskEntity.setUpdateUsername(loggedUserInfo.username());
 
         taskRepository.save(taskEntity);
 
@@ -95,7 +107,17 @@ public class TaskService {
     }
 
     @Transactional
-    public void deleteTask(Long taskId) {
+    public void deleteTask(Long taskId, LoggedUserInfo loggedUserInfo) {
+        Long creatorUserId = taskRepository.getCreatorUserId(taskId);
+
+        if (!canModifyTask(creatorUserId, loggedUserInfo)) {
+            throw new RuntimeException("You are not allowed to delete this task"); // TODO: custom exception
+        }
         taskRepository.deleteById(taskId);
+    }
+
+    private boolean canModifyTask(Long creatorUserId, LoggedUserInfo loggedUserInfo) {
+        // TODO: who can update a task? The creator, the assignee, the admin? -> let's assume the creator and the admin
+        return loggedUserInfo.isAdmin() || creatorUserId.equals(loggedUserInfo.userId());
     }
 }
